moved

+++ v4_file_per_expression/bin/generate_new_random_index.sh
@@ -2,7 +2,9 @@
 # Physics Derivation Graph
 # 20160129
 
-rm temp_id.dat
+if [ -f temp_id.dat ] ; then
+  rm temp_id.dat
+fi
 find ../derivations -depth 1 -type d | while read this_dir; do
     #echo "looking for temporary IDs in ${this_dir}"
     cat "${this_dir}/derivation_edge_list.csv" | tr ',' '\n' >> temp_id.dat # convert from two column to single column
@@ -15,8 +17,9 @@ mv temp_id2.dat temp_id.dat
 current_guess=`head -c7 <(echo $RANDOM$RANDOM$RANDOM$RANDOM$RANDOM)`
 if grep -Fxq $current_guess temp_id.dat; then echo "already in file; try again"; else echo "temp ID not in use: $current_guess"; fi
 
-
-rm expr_id.dat
+if [ -f expr_id.dat ]; then
+  rm expr_id.dat
+fi
 find ../derivations -depth 1 -type d | while read this_dir; do
     #echo "looking for expression IDs in ${this_dir}"
     cat "${this_dir}/expression_identifiers.csv" | cut -d',' -f2 >> expr_id.dat # convert from two column to single column
